SCMOD-7044: Use HTTP data store in integration tests
I missed this from my previous PR where I updated the build image only: #39

The previous PR builds on linux because the integration tests just create an instance of a FileSystemDataStore (implicitly via the testing framework), but don't use it for anything. Although this previous PR works and this PR is not strictly necessary, I thought it best to update the intergration tests anyway to use the HTTPDataStore instead of the FileSystemDataStore data store in case the integration tests were updated in the future to start to use a data store.